Update gem structure
Adds Form as a top-level object to handle loading from JSON
or loading from API.

Structures existing data models into FormData namespace to
begin separation for allowing Websocket data and additional
gem functionality.
