+++ lib/typed_form.rb
@@ -1,11 +1,12 @@
 require "httparty"
 require "forwardable"
 require "arendelle"
-require "typed_form/client"
-require "typed_form/json_response_handler"
-require "typed_form/form_response"
-require "typed_form/question"
-require "typed_form/form_answers"
+require "typed_form/api/client"
+require "typed_form/parsed_json"
+require "typed_form/form_data/question"
+require "typed_form/form_data/answers"
+require "typed_form/form_data/form_submission"
+require "typed_form/form"
 require "typed_form/version"
 
 module TypedForm
