+++ lib/typed_form/parsed_json.rb
@@ -1,13 +1,15 @@
 module TypedForm
-  class JSONResponseHandler
+  class ParsedJson
     extend Forwardable
     attr_reader :json
 
     def_delegators :parsed_json, :questions, :responses
 
-    def initialize(json)
-      @json = json
+    def initialize(json:)
+      @json = json.freeze
     end
+
+    private
 
     def parsed_json
       @_parsed_json ||= JSON.parse(json, object_class: Arendelle)
