--- spec/form_answers_spec.rb
@@ -1,4 +1,0 @@
-require "spec_helper"
-
-RSpec.describe TypedForm::FormAnswers do
-end
